Adds funnel chart visualization & storybook demo
Funnel chart viz based on github.com/jakezatecky/d3-funnel. Viz
has been tested both with Storybook and the Superset Legacy API
(viz.py), functional for both. Ready to be integrated into
Apache Superset as a working plugin.

Funnel charts are a type of chart, often used to represent stages
in a sales process and show the amount of potential revenue for
each stage. This type of chart can also be useful in identifying
potential problem areas in an organization’s sales processes.